--- src/main/java/scs/planus/dto/todo/CategoryChangeRequestDto.java
@@ -1,29 +1,0 @@
-package scs.planus.dto.todo;
-
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import scs.planus.domain.Color;
-import scs.planus.domain.TodoCategory;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class CategoryChangeRequestDto {
-    private String name;
-    private Color color;
-
-    @Builder
-    public CategoryChangeRequestDto(String name, Color color) {
-        this.name = name;
-        this.color = color;
-    }
-
-    public TodoCategory toEntity() {
-        return TodoCategory.builder()
-                .name(this.name)
-                .color(this.color)
-                .build();
-    }
-
-}
